GencGuney/requirements.txt
gdown
lmdb
numpy
pillow
requests
torch>=2.0.0
torchvision>=0.15.0
tqdm
